requirements.txt
# Python Dependencies for AI-Powered E-Commerce Semantic Search

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# OpenAI for LLM capabilities
openai>=1.0.0

# Embeddings and Re-ranking
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Search
faiss-cpu>=1.7.4
numpy>=1.24.0

# Environment Configuration
python-dotenv>=1.0.0

# Logging and Monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
httpx>=0.25.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
